% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nft_upload.R
\name{nft_CertUpload}
\alias{nft_CertUpload}
\title{上传证书至数据库}
\usage{
nft_CertUpload(
  conn = tsda::conn_rds("nftns"),
  FFileName,
  FTxtId,
  FFileHash,
  blockid,
  FCertUrl
)
}
\arguments{
\item{conn}{连接}

\item{FFileName}{文件名}

\item{FTxtId}{文件ID}

\item{FCompanyPrefix}{公司前缀}
}
\value{
返回值
}
\description{
上传证书至数据库
}
\examples{
nft_CertUpload()
}
