% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nft_upload.R
\name{nft_fileUpload}
\alias{nft_fileUpload}
\title{上传文件至数据库}
\usage{
nft_fileUpload(
  conn = tsda::conn_rds("nftns"),
  FCompanyPrefix,
  FFileName,
  FTxtId
)
}
\arguments{
\item{conn}{连接}

\item{FCompanyPrefix}{公司前缀}

\item{FFileName}{文件名}

\item{FTxtId}{文件ID}
}
\value{
返回值
}
\description{
上传文件至数据库
}
\examples{
nft_fileUpload()
}
