% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nft_upload.R
\name{nft_queryList}
\alias{nft_queryList}
\title{查询证书列表}
\usage{
nft_queryList(conn = tsda::conn_rds("nftns"))
}
\arguments{
\item{conn}{查询}
}
\value{
返回值
}
\description{
查询证书列表
}
\examples{
nft_queryList()
}
